Refactored Drawable helper to Uno.UWP
